Rename timetool python package to lcls2_timetool
