+++ firmware/common/python/lcls2_timetool/_TimeToolKcu1500.py
@@ -3,14 +3,14 @@ import pyrogue as pr
 import axipcie
 import lcls2_pgp_fw_lib.hardware.XilinxKcu1500
 
-import timetool
+import lcls2_timetool
 
 class TimeToolKcu1500(pr.Device):
     def __init__(self, numLanes=1, pgp3=False, **kwargs):
         super().__init__(**kwargs)
 
         # The time tool application
-        self.add(timetool.Application(
+        self.add(lcls2_timetool.Application(
             offset = 0x00C0_0000,
             numLanes = numLanes))
