--- firmware/common/python/timetool/__init__.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-
-from timetool._Application           import *
-from timetool._AppLane               import *
-from timetool._Fex                   import *
-from timetool._Prescale              import *
-from timetool._ByPass                import *
-from timetool._FIR                   import *
-from timetool._FrameIIR              import *
-from timetool._FrameSubtractor       import *
-from timetool._Piranha4VcsEmu        import *
-from timetool._TimeToolKcu1500       import *
-from timetool._RunControl            import *
-from timetool._TimeToolKcu1500Root   import *
